+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/action/AbstractTapChangerAction.java
@@ -1,0 +1,36 @@
+/**
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+package com.powsybl.security.action;
+
+import com.powsybl.iidm.network.ThreeWindingsTransformer;
+
+import java.util.Objects;
+import java.util.Optional;
+
+/**
+ * @author Etienne Lesot <etienne.lesot@rte-france.com>
+ */
+public abstract class AbstractTapChangerAction extends AbstractAction {
+
+    private final String transformerId;
+    private final ThreeWindingsTransformer.Side side;
+
+    protected AbstractTapChangerAction(String id, String transformerId, ThreeWindingsTransformer.Side side) {
+        super(id);
+        this.transformerId = Objects.requireNonNull(transformerId);
+        this.side = side;
+    }
+
+    public String getTransformerId() {
+        return transformerId;
+    }
+
+    public Optional<ThreeWindingsTransformer.Side> getSide() {
+        return Optional.ofNullable(side);
+    }
+}
